fix: propagate model dtype to all zero-tensor initialisations
Every torch.zeros that feeds into an RNN hidden state, is
concatenated with model outputs, or participates in arithmetic
with bfloat16 tensors was missing an explicit dtype= argument,
defaulting to float32 and crashing at runtime.

Affected sites:
- self_model.py: init_state h (direct GRU input) — primary crash
- htm_layer.py: h_prev list for GRUCell steps
- fast_weight.py: W_fast initial matrix
- insula.py: effort_proxy padding cat'd with nt_vec
- neural_geometry.py: accumulator in encode_sequence
- claustrum.py: modality padding in flat_parts cat
- theory_of_mind.py: cf_resp, social_error/reward, _disabled_output, entity_context_vector
- learned_opt.py: LSTM h/c in _get_state
- brain.py: reward_proxy_for_amyg, emotional_valence, arousal,
  claustrum_out salience, geo_out scalars, insula_salience,
  cereb_error, mesolimbic zero reward — all now use the dtype
  variable already derived at the top of forward_lm

+++ neuroslm/modules/claustrum.py
@@ -134,7 +134,7 @@ class Claustrum(nn.Module):
         # Pad if fewer modalities than expected
         flat_parts = [p for p in projected]
         while len(flat_parts) < self.n_modalities:
-            flat_parts.append(torch.zeros(B, self.d_sem, device=device))
+            flat_parts.append(torch.zeros(B, self.d_sem, device=device, dtype=projected[0].dtype))
         flat = torch.cat(flat_parts[:self.n_modalities], dim=-1)
         cell_out = self.cells(flat)  # (B, D)
